Fase 1, módulo 2: persistência com trava contra divergência de esquema
Modelos SQLAlchemy, repositórios e 35 testes em SQLite em memória. Destrava os
critérios 19 (auditoria de parâmetros) e parte do 20 (LGPD no modelo de dados).

Duas fontes de verdade eram o maior risco deste módulo. Os modelos são a fonte
do runtime e banco/schema.sql segue como DDL PostgreSQL comentado de
referência; test_schema_sql_nao_divergiu_dos_modelos compara tabela a tabela e
coluna a coluna. Sem essa trava, uma coluna acrescentada ao modelo e esquecida
no SQL só apareceria na migração de produção, meses depois.

O que o banco garante sozinho, com teste para cada um
- concurso duplicado é recusado mesmo burlando o repositório (corrida entre
  workers não passa)
- backtest sem semente é recusado: backtest que não se reproduz não se audita
- partição fora de ordem cronológica é recusada pelo CHECK
- janela de walk-forward não pode invadir o treino
- alteração de estratégia cria versão nova e não toca a anterior
- apagar usuário leva estratégias, versões e jogos

LGPD no modelo, também testado: só e-mail e hash — sem nome, CPF, telefone,
endereço ou data de nascimento; consentimentos datados; ip_hash em vez de IP;
Auditoria expõe apenas registrar() e listar(), porque auditoria reescrevível
não é auditoria.

Portabilidade: ARRAY(SMALLINT)/JSONB no PostgreSQL e JSON no SQLite via
with_variant, para a suíte rodar sem banco no ar. A chave primária de lotteries
passou de SMALLINT para INTEGER porque o SQLite só autoincrementa INTEGER
PRIMARY KEY — tipo que só funciona em um dos dois bancos não é portátil.

Integração real coberta: o histórico que sai do repositório entra no motor de
backtest sem adaptação.

172 testes, todos passando, sem rede.

+++ analytics/banco/schema.sql
@@ -19,7 +19,7 @@ CREATE EXTENSION IF NOT EXISTS pgcrypto;   -- gen_random_uuid, crypt
 -- ===========================================================================
 
 CREATE TABLE lotteries (
-    id               SMALLSERIAL PRIMARY KEY,
+    id               SERIAL      PRIMARY KEY,
     code             TEXT        NOT NULL UNIQUE,
     name             TEXT        NOT NULL,
     -- 'dezenas' | 'composta' (dezenas + complemento) | 'colunas' (Super Sete)
@@ -41,7 +41,7 @@ CREATE TABLE lotteries (
 -- concurso de 2019 infla o custo e produz um ROI falso no histórico inteiro.
 CREATE TABLE price_history (
     id            BIGSERIAL PRIMARY KEY,
-    lottery_id    SMALLINT NOT NULL REFERENCES lotteries(id) ON DELETE CASCADE,
+    lottery_id    INTEGER  NOT NULL REFERENCES lotteries(id) ON DELETE CASCADE,
     valid_from    DATE     NOT NULL,
     bet_price     NUMERIC(10, 2) NOT NULL CHECK (bet_price > 0),
     source        TEXT     NOT NULL DEFAULT 'caixa',
@@ -54,7 +54,7 @@ CREATE TABLE price_history (
 
 CREATE TABLE draws (
     id              BIGSERIAL PRIMARY KEY,
-    lottery_id      SMALLINT   NOT NULL REFERENCES lotteries(id) ON DELETE CASCADE,
+    lottery_id      INTEGER    NOT NULL REFERENCES lotteries(id) ON DELETE CASCADE,
     contest_number  INTEGER    NOT NULL CHECK (contest_number > 0),
     draw_date       DATE       NOT NULL,
     -- Array preserva a ORDEM: o Super Sete depende dela, e ordenar destruiria
@@ -142,7 +142,7 @@ CREATE TABLE generated_games (
     id                  BIGSERIAL PRIMARY KEY,
     user_id             UUID     NOT NULL REFERENCES users(id) ON DELETE CASCADE,
     strategy_version_id BIGINT   REFERENCES strategy_versions(id) ON DELETE SET NULL,
-    lottery_id          SMALLINT NOT NULL REFERENCES lotteries(id),
+    lottery_id          INTEGER  NOT NULL REFERENCES lotteries(id),
     numbers             SMALLINT[] NOT NULL,
     generated_at        TIMESTAMPTZ NOT NULL DEFAULT now(),
     cost                NUMERIC(10, 2) NOT NULL CHECK (cost >= 0),
@@ -159,7 +159,7 @@ CREATE INDEX generated_games_por_usuario ON generated_games (user_id, generated_
 CREATE TABLE backtests (
     id                  BIGSERIAL PRIMARY KEY,
     strategy_version_id BIGINT   NOT NULL REFERENCES strategy_versions(id) ON DELETE CASCADE,
-    lottery_id          SMALLINT NOT NULL REFERENCES lotteries(id),
+    lottery_id          INTEGER  NOT NULL REFERENCES lotteries(id),
 
     -- A partição é gravada como faixas de concurso, para ser reconstituível
     -- meses depois sem depender de percentuais nem do tamanho do histórico
@@ -233,7 +233,7 @@ CREATE TABLE backtest_windows (
 
 CREATE TABLE import_runs (
     id            BIGSERIAL PRIMARY KEY,
-    lottery_id    SMALLINT NOT NULL REFERENCES lotteries(id),
+    lottery_id    INTEGER  NOT NULL REFERENCES lotteries(id),
     started_at    TIMESTAMPTZ NOT NULL DEFAULT now(),
     finished_at   TIMESTAMPTZ,
     source        TEXT NOT NULL,
